drag and drop effect as a separate component
